[ait] moved ait into separate folder
