+++ src/goto-programs/abstract-interpretation/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(abstract-interpretation ai.cpp ai_domain.cpp interval_domain.cpp interval_analysis.cpp)
+target_link_libraries(abstract-interpretation fmt::fmt)
