+++ src/goto-programs/abstract-interpretation/ai.h
@@ -7,7 +7,7 @@
 #include <iosfwd>
 #include <map>
 #include <memory>
-#include <goto-programs/ai_domain.h>
+#include <goto-programs/abstract-interpretation/ai_domain.h>
 #include <goto-programs/goto_functions.h>
 #include <util/xml.h>
 #include <util/expr.h>
