+++ src/goto-programs/abstract-interpretation/interval_template.h
